requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlmodel==0.0.14
httpx==0.25.2
python-dotenv==1.0.1
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-httpx==0.25.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic-settings==2.1.0
jinja2==3.1.3
python-multipart